Rename v1/resource dir to v1/crd
