+++ pkg/scaffold/v1/crd/crd_sample.go
@@ -14,7 +14,7 @@ See the License for the specific language governing permissions and
 limitations under the License.
 */
 
-package resource
+package crd
 
 import (
 	"fmt"
