+++ pkg/scaffold/v1/crd/types.go
@@ -14,7 +14,7 @@ See the License for the specific language governing permissions and
 limitations under the License.
 */
 
-package resource
+package crd
 
 import (
 	"fmt"
